Script de desinstalacion terminado

+++ inst-uinst/uninstall.sh
@@ -1,4 +1,41 @@
 #!/bin/bash
-sudo apt-get -y install python3-tk
-sudo apt-get -y install python3-pip
-sudo apt-get -y install python3-venv
+
+# begin
+echo ""
+
+# tkinter
+
+echo "Would you like to delete tkinter GUI environment of your system? (y/n): "
+read -n 1 input
+
+if [ "$input" = "y" ]; then
+
+	sudo apt-get -y install python3-tk
+
+fi
+
+# pip
+
+echo -e "\nWould you like to delete pip package of your system? (y/n): "
+read -n 1 input
+
+if [ "$input" = "y" ]; then
+
+	sudo apt-get -y install python3-pip
+
+fi
+
+# venv
+
+echo -e "\nWould you like to delete venv package of your system? (y/n): "
+read -n 1 input
+
+if [ "$input" = "y" ]; then
+
+	sudo apt-get -y install python3-venv
+	
+fi
+
+
+# end
+echo -e "\n"
